backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
starlette==0.27.0

# Authentication/Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Database
motor==3.3.1
pymongo==4.5.0
redis==5.0.1
pinecone-client==2.2.4

# Environment variables
python-dotenv==1.0.0

# AI/NLP
langchain==0.0.305
openai==1.2.4
anthropic==0.8.0

# Utilities
tenacity==8.2.3
pytz==2023.3
python-dateutil==2.8.2
